requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1

sqlalchemy==2.0.32
asyncpg==0.29.0
alembic==1.13.2

pydantic==2.7.4
pydantic-settings==2.4.0
email-validator==2.2.0

python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9

httpx==0.27.0

pytest==8.3.2
pytest-asyncio==0.23.8

redis==5.0.8